+++ R/nlmixrAddObjectiveFunction.R
@@ -3,7 +3,7 @@
 #' @param fit nlmixr fit object
 #' @param objDf nlmixr objective function data frame which has column
 #'   names "OBJF", "AIC", "BIC", "Log-likelihood" and
-#'   "Condition Number"
+#'   "Condition#(Cov)" "Condition#(Cor)"
 #' @param type Objective Function Type
 #' @param etaObf Eta objective function table to add (with focei) to
 #'   give focei objective function
@@ -27,9 +27,15 @@ nlmixrAddObjectiveFunctionDataFrame <- function(fit, objDf, type, etaObf=NULL) {
     } else if (!is.na(.inRow[[2]])) {
       .cn <- .inRow[[2]]
     }
+    .cnr <- NA_real_
+    if (!is.na(.inRow2[[3]])) {
+      .cnr <- .inRow2[[3]]
+    } else if (!is.na(.inRow[[3]])) {
+      .cnr <- .inRow[[3]]
+    }
     if (is.na(.inRow2[[1]][[1]])) {
       # Here the original data frame is NA, that is the objective function has not been calculated
-      .tmp <- cbind(.inRow[[1]], data.frame("Condition Number"=.cn, check.names=FALSE))
+      .tmp <- cbind(.inRow[[1]], data.frame("Condition#(Cov)"=.cn, "Condition#(Cor)"=.cnr, check.names=FALSE))
       row.names(.tmp) <- type
       assign("objDf", .tmp, envir=fit$env)
       setOfv(fit, type)
@@ -37,7 +43,8 @@ nlmixrAddObjectiveFunctionDataFrame <- function(fit, objDf, type, etaObf=NULL) {
       if (any(.rownames == type)) stop("objective function '", type, "' already present", call.=FALSE)
       # Now the original data frame is not NA.
       .tmp <- rbind(.inRow[[1]], .inRow2[[1]])
-      .tmp[["Condition Number"]] <- .cn
+      .tmp[["Condition#(Cov)"]] <- .cn
+      .tmp[["Condition#(Cor)"]] <- .cnr
       row.names(.tmp) <- c(type, .rownames)
       assign("objDf", .tmp, envir=fit$env)
       setOfv(fit, type)
@@ -45,14 +52,23 @@ nlmixrAddObjectiveFunctionDataFrame <- function(fit, objDf, type, etaObf=NULL) {
   } else {
     if (any(.rownames == type)) stop("objective function '", type, "' already present", call.=FALSE)
     ## Now there is at least one interesting objective function
-    .cn <- .cur[["Condition Number"]][1]
+    .cn <- .cur[["Condition#(Cov)"]][1]
     if (is.null(.cn)) {
       .cn <- NA_real_
     }
     if (is.na(.cn) & !is.na(.inRow[[2]])) {
       .cn <- .inRow[[2]]
     }
-    .cur <- rbind(.cur[, names(.cur) != "Condition Number"], .inRow[[1]])
+    .cnr <- .cur[["Condition#(Cor)"]][1]
+    if (is.null(.cnr)) {
+      .cnr <- NA_real_
+    }
+    if (is.na(.cnr) & !is.na(.inRow[[3]])) {
+      .cnr <- .inRow[[3]]
+    }
+    .cur <- rbind(.cur[, !(names(.cur) %in% c("Condition#(Cor)", "Condition#(Cov)"))], .inRow[[1]])
+    .cur[["Condition#(Cov)"]] <- .cn
+    .cur[["Condition#(Cor)"]] <- .cnr
     row.names(.cur) <- c(.rownames, type)
     assign("objDf", .cur, envir=fit$env)
     setOfv(fit, type)