+++ src/inner.cpp
@@ -5877,13 +5877,18 @@ void foceiFinalizeTables(Environment e){
   }
   LogicalVector skipCov = e["skipCov"];
 
-  if (covExists && op_focei.eigen){
+  if (covExists) {
+    Function loadNamespace("loadNamespace", R_BaseNamespace);    
+    Environment nlmixr2 = loadNamespace("nlmixr2est");
+    Function getCor = nlmixr2[".cov2cor"];
+    e["fullCor"] = getCor(e["cov"]);
+    arma::mat cor = as<arma::mat>(e["fullCor"]);
+    cor.diag().ones();
     arma::vec eigval;
     arma::mat eigvec;
-
-    eig_sym(eigval, eigvec, cov);
-    e["eigen"] = eigval;
-    e["eigenVec"] = eigvec;
+    eig_sym(eigval, eigvec, cor);
+    e["eigenCor"] = eigval;
+    e["eigenVecCor"] = eigvec;
     unsigned int k=0;
     if (eigval.size() > 0){
       double mx=std::fabs(eigval[0]), mn, cur;
@@ -5897,9 +5902,35 @@ void foceiFinalizeTables(Environment e){
           mn=cur;
         }
       }
-      e["conditionNumber"] = mx/mn;
+      e["conditionNumberCor"] = mx/mn;
     } else {
-      e["conditionNumber"] = NA_REAL;
+      e["conditionNumberCor"] = NA_REAL;
+    }
+  }
+
+  if (covExists && op_focei.eigen){
+    arma::vec eigval;
+    arma::mat eigvec;
+
+    eig_sym(eigval, eigvec, cov);
+    e["eigenCov"] = eigval;
+    e["eigenVecCov"] = eigvec;
+    unsigned int k=0;
+    if (eigval.size() > 0){
+      double mx=std::fabs(eigval[0]), mn, cur;
+      mn=mx;
+      for (k = eigval.size(); k--;){
+        cur = std::fabs(eigval[k]);
+        if (cur > mx){
+          mx=cur;
+        }
+        if (cur < mn){
+          mn=cur;
+        }
+      }
+      e["conditionNumberCov"] = mx/mn;
+    } else {
+      e["conditionNumberCov"] = NA_REAL;
     }
   }
   arma::vec se1;
@@ -6255,10 +6286,11 @@ void foceiFinalizeTables(Environment e){
     }
   }
   List objDf;
-  if (e.exists("conditionNumber")){
+  if (e.exists("conditionNumberCov")){
     objDf = List::create(_["OBJF"] = as<double>(e["objective"]), _["AIC"]=as<double>(e["AIC"]),
                          _["BIC"] = as<double>(e["BIC"]), _["Log-likelihood"]=as<double>(e["logLik"]),
-                         _["Condition Number"]=as<double>(e["conditionNumber"]));
+                         _["Condition#(Cov)"]=as<double>(e["conditionNumberCov"]),
+                         _["Condition#(Cor)"]=as<double>(e["conditionNumberCor"]));
   } else {
     objDf = List::create(_["OBJF"] = as<double>(e["objective"]), _["AIC"]=as<double>(e["AIC"]),
                          _["BIC"] = as<double>(e["BIC"]), _["Log-likelihood"]=as<double>(e["logLik"]));